+++ Sources/IMsgCore/IMsgBridgeProtocol.swift
@@ -31,8 +31,8 @@ public enum IMsgBridgeProtocol {
 
   public static func defaultResponseTimeout(for action: BridgeAction) -> TimeInterval {
     switch action {
-    case .sendMessage, .sendMultipart, .sendAttachment, .sendPoll, .sendReaction,
-      .createChat:
+    case .sendMessage, .sendMultipart, .sendAttachment, .sendPoll, .sendPollVote,
+      .sendReaction, .createChat:
       return defaultSendResponseTimeout
     default:
       return defaultResponseTimeout
@@ -65,6 +65,7 @@ public enum BridgeAction: String, Sendable, CaseIterable {
   case sendMultipart = "send-multipart"
   case sendAttachment = "send-attachment"
   case sendPoll = "send-poll"
+  case sendPollVote = "send-poll-vote"
   case sendReaction = "send-reaction"
   case notifyAnyways = "notify-anyways"